# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Lotofond API'
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /send/application:
    post:
      parameters:
        - name: Content-Language
          in: header
          description: Заголовок для определения локали (ru, en, zh_CN)
          required: true
      tags:
        - Applications
      summary: Отправка заявки на покупку без ЕЦП/через агента на почту менеджеру
      operationId: sendApplication
      requestBody:
        description: данные для отправления заявки (если заявка на покупку через агента - не указываем lotId и указываем name, если на покупку без ЕЦП - наоборот)
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - phone
                - paymentTradingTypes
              properties:
                email:
                  type: string
                lotId:
                  type: integer
                name:
                  type: string
                phone:
                  type: string
                paymentTradingTypes:
                  type: array
                  items:
                    type: string
                    enum:
                      - purchaseBidByAgent
                      - infoAboutLot
                      - consultation,
                      - accompanimentFAS
                socialsForAnswer:
                  type: array
                  items:
                    type: string
                    enum:
                      - Vk
                      - Viber
                      - WhatsApp
                      - Telegram
                dateForCallback:
                  type: string
      responses:
        "200":
          description: заявка успешно отправлена
        "422":
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
  /send/purchase/instructions:
    post:
      parameters:
        - name: Content-Language
          in: header
          description: Заголовок для определения локали (ru, en, zh_CN)
          required: true
      tags:
        - Applications
      summary: Оплата услуги "Покупка без ЭЦП - инструкция"
      operationId: purchaseInstructions
      requestBody:
        description: данные для отправления заявки
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - lotId
              properties:
                email:
                  type: string
                lotId:
                  type: integer
      responses:
        "200":
          description: Создание платежа прошло успешно, в ответе ссылка для его подтверждения
          content:
            application/json:
              schema:
                type: object
                properties:
                  redirectUrl:
                    type: string
              example:
                redirectUrl: https://paymaster.ru/cpay/7ec1c57594b54d059713003128b60286
        403:
          description: при создании платежа paymaster вернул ошибку
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "422":
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
  /send/receipt/egrn:
    post:
      parameters:
        - name: Content-Language
          in: header
          description: Заголовок для определения локали (ru, en, zh_CN)
          required: true
      tags:
        - Applications
      summary: Оплата услуги "Выписка ЕГРН"
      operationId: receiptEgrn
      requestBody:
        description: данные для отправления заявки
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cadastralNumber
                - lotId
                - email
              properties:
                lotId:
                  type: integer
                email:
                  type: string
                cadastralNumber:
                  type: string
                  example: 61:44:0071802:1786
                answerFormat:
                  type: string
                  enum:
                    - pdf
                    - html
                    - zip
                  description: формат файла, в котором пользователю прислать выписку. По умолчанию - pdf
      responses:
        "200":
          description: Создание платежа прошло успешно, в ответе ссылка для его подтверждения
          content:
            application/json:
              schema:
                type: object
                properties:
                  redirectUrl:
                    type: string
              example:
                redirectUrl: https://paymaster.ru/cpay/7ec1c57594b54d059713003128b60286
        403:
          description: при создании платежа paymaster вернул ошибку
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "422":
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
  /send/question:
    post:
      parameters:
        - name: Content-Language
          in: header
          description: Заголовок для определения локали (ru, en, zh_CN)
          required: true
      tags:
        - Applications
      summary: Отправка вопроса на почту менеджеру
      operationId: sendQuestion
      requestBody:
        description: данные для отправления вопроса
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - email
                - topic
                - question
              properties:
                email:
                  type: string
                topic:
                  type: string
                question:
                  type: string
                files[]:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: вопрос успешно отправлен
        "422":
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
  /send/contacts:
    post:
      parameters:
        - name: Content-Language
          in: header
          description: Заголовок для определения локали (ru, en, zh_CN)
          required: true
      tags:
        - Applications
      summary: Отправка вопроса на почту менеджеру
      operationId: sendContacts
      requestBody:
        description: данные для отправления вопроса
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question
                - communicationType
                - communication
              properties:
                question:
                  type: string
                communication:
                  type: string
                communicationType:
                  type: string
                  enum:
                    - email
                    - phone
      responses:
        "200":
          description: заявка успешно отправлена
        "422":
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /payment:
    post:
      tags:
        - Payments
      summary: запрос на создание платежа
      operationId: payment
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                tariffId:
                  type: integer
                  format: int64
                  description: Айдишник тарифа
              example:
                tariffId: 1
      responses:
        "200":
          description: Создание платежа прошло успешно, в ответе ссылка для его подтверждения
          content:
            application/json:
              schema:
                type: object
                properties:
                  redirectUrl:
                    type: string
              example:
                redirectUrl: https://paymaster.ru/cpay/7ec1c57594b54d059713003128b60286
        403:
          description: при создании платежа paymaster вернул ошибку
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    Error:
      required:
        - code
        - detail
        - title
      type: object
      properties:
        code:
          type: integer
          format: int32
        title:
          type: string
        detail:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT